aipy_src/_miriad: fix mask I/O on large data sets
I was stuck on this because my first attempt broke things subtly. It turned
out that this was because I tried expressing the mask `length` and `size`
fields as size_t's. Some of the mask I/O code adds `offset` and `length` with
intended result being a negative number, and apparently C upcasts the
expression to unsigned without complaint. So that broke things. Expressing
`length` and `size` as off64_t's seems to make things work. (Note that off64_t
is a signed type.)

Fix verified on a 30 GB dataset I happen to have generated recently.

+++ aipy_src/_miriad/mir/miriad.h
@@ -279,8 +279,8 @@ void xysetpl_c (int tno, int naxis, Const int *axes);
 /* maskio.c */
 char *mkopen_c (int tno, char *name, char *status);
 void mkclose_c (char *handle);
-int  mkread_c  (char *handle, int mode, int *flags, int offset, int n, int nsize);
-void mkwrite_c (char *handle, int mode, int *flags, int offset, int n, int nsize);
+int  mkread_c  (char *handle, int mode, int *flags, off64_t offset, int n, int nsize);
+void mkwrite_c (char *handle, int mode, int *flags, off64_t offset, int n, int nsize);
 void mkflush_c (char *handle);
